Allow overriding of coercers in the registry in a local context
Sometimes transformation away from the spec'ed format into a similar,
but slightly different format is useful. For example when dealing with
native datatypes that a storage engine doesn't understand. In these
circumstances it's useful to be able to leverage the spec structure,
but apply a different transformation for a set of keys.

By adding the dynamic var *overrides* we are able to rebind the
coercer for any key in a limited scope, allowing us to perform custom
transformations without having to deal with structure.

This differs from coerce-structure's override functionality in that
it allows coercers to be overridden by qualified key, with the new
transformation being used anywhere that key appears.